refactor(telemetry-settings): move definition from main to api package
related to https://github.com/podman-desktop/podman-desktop/issues/14361
